added functionality for Ulam-like triangle spiral
no automatic plotting yet, but optional arguments allow for
spirals derived from arbitrary triangles.
Arguments ask for 3 2d vectors; if these add to [0,0] a spiral will be made
whose boundary approaches the triangle made by stacking those tip-to-tail.

The math for plotting assumes these vectors add to [0,0]. Behavior has not
been explored where this is not the case.

TODO: add better documentation, comment out debug prints, add auto plotting

Next: regular n-gon

I wanted to eventually allow the user to specify a more arbitrary pattern
of when to extend the segment and how far to rotate for the next segment,
but I haven't yet figured out an efficient way to calculate the position of
a point without iterating through the spiral up to that point, and I think I'd
rather focus on polygons and add features for users to play around with those.